2017/12/6  9.02

+++ app/src/main/java/com/pilipa/fapiaobao/ui/DemandActivity.java
@@ -638,7 +638,7 @@ public class DemandActivity extends BaseActivity{
         WindowManager.LayoutParams lp = dialogWindow.getAttributes(); // 获取对话框当前的参数值
         lp.x = 0; // 新位置X坐标
         lp.y = 0; // 新位置Y坐标
-        lp.width = (int) getResources().getDisplayMetrics().widthPixels; // 宽度
+        lp.width = getResources().getDisplayMetrics().widthPixels; // 宽度
         root.measure(0, 0);
         lp.height = root.getMeasuredHeight();
